requirements.txt
streamlit==1.32.0
folium==0.14.0
streamlit-folium==0.15.0
google-api-python-client==2.107.0
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pyperclip==1.8.2
geocoder==1.38.1
pandas==2.1.1

